feat: connect scap to tauri app state
